# docs/swagger.yaml
definitions:
  dto.CreateTodoRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  dto.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dto.UpdateTodoRequest:
    properties:
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  models.Todo:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Email va parol bilan kirish, token qaytaradi
      parameters:
      - description: Email va parol
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Foydalanuvchi kirishi (Sign In)
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Yangi foydalanuvchini bazaga qo'shadi va token qaytaradi
      parameters:
      - description: Foydalanuvchi ma'lumotlari
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Foydalanuvchi yaratish (Sign Up)
      tags:
      - Auth
  /todos:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      security:
      - BearerAuth: []
      summary: Barcha todolar
      tags:
      - Todos
  /todos/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Todo ma'lumotlari
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTodoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Yaratildi
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Todo yaratish
      tags:
      - Todos
  /todos/delete:
    delete:
      parameters:
      - description: Todo ID
        in: query
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: O'chirildi
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Todo o'chirish
      tags:
      - Todos
  /todos/get:
    get:
      parameters:
      - description: Todo ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      security:
      - BearerAuth: []
      summary: Bitta todo
      tags:
      - Todos
  /todos/update:
    put:
      consumes:
      - application/json
      parameters:
      - description: Todo ID
        in: query
        name: id
        required: true
        type: integer
      - description: Todo ma'lumotlari
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTodoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Yangilandi
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Todo yangilash
      tags:
      - Todos
swagger: "2.0"